+++ mlc/minilang_parser.ml
@@ -4177,7 +4177,7 @@ function _compile_predefined_values()
     CompileValue("TARGET_ABI", _compile_target_abi),
     CompileValue("TARGET_FORMAT", _compile_target_format),
     CompileValue("POINTER_SIZE", 8),
-    CompileValue("MINILANG_VERSION", "1.2.1")
+    CompileValue("MINILANG_VERSION", "1.2.2")
   ]
 end function